Split periscope_ui into smaller modules
